requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-ai-documentintelligence==1.0.0b4
azure-core==1.32.0
cachetools==5.4.0
certifi==2024.6.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
cssselect==1.2.0
cssutils==2.11.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isodate==0.7.2
Jinja2==3.1.4
jiter==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.19
langchain-openai==0.2.9
langsmith==0.1.144
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
numpy==1.26.4
openai==1.55.0
orjson==3.10.5
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
premailer==3.10.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyMuPDF==1.24.13
pypdfium2==4.30.0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==1.0.2
PyYAML==6.0.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==0.22.0
websockets==12.0
XlsxWriter==3.2.0
yagmail==0.15.293
